+++ mek_data_class_generator/lib/builder.dart
@@ -1,5 +1,4 @@
 import 'package:build/build.dart';
-import 'package:dart_style/dart_style.dart';
 import 'package:mek_data_class_generator/mek_data_class_generator.dart';
 import 'package:mek_data_class_generator/src/configs.dart';
 import 'package:source_gen/source_gen.dart';
@@ -8,12 +7,5 @@ import 'package:source_gen/source_gen.dart';
 Builder dataClass(BuilderOptions options) {
   final config = Config.fromJson(options.config);
 
-  return SharedPartBuilder(
-    [DataClassGenerator(config: config)],
-    'data_class',
-    formatOutput: DartFormatter(
-      pageWidth: config.pageWidth,
-      fixes: {StyleFix.singleCascadeStatements},
-    ).format,
-  );
+  return SharedPartBuilder([DataClassGenerator(config: config)], 'data_class');
 }